Flask can take in parameters from GET request
